@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using HybridDb.Config;
 using ShouldBeLike;
 using Shouldly;
